printf

+++ ft_putnbr_base.c
@@ -12,9 +12,9 @@
 
 #include "ft_printf.h"
 
-int	ft_putnbr_base(unsigned long long nbr, char *base, int base_len, int *count)
+int	ft_putnbr_base(unsigned long int nbr, char *base, int base_len, int *count)
 {
-	if (nbr >= (unsigned long long)base_len)
+	if (nbr >= (unsigned long int)base_len)
 	{
 		if (ft_putnbr_base(nbr / base_len, base, base_len, count) == -1)
 			return (-1);
